refactor : DatabasePoolManager에 close 추가 및 Dao 코드 수정
