+++ nest_server_0616/src/org/kosa/nest/common/DatabasePoolManager.java
@@ -1,5 +1,10 @@
 package org.kosa.nest.common;
 
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
 import javax.sql.DataSource;
 
 import org.apache.commons.dbcp.BasicDataSource;
@@ -28,4 +33,17 @@ public class DatabasePoolManager {
     public DataSource getDataSource() {
         return dataSource;
     }
+    
+    public static void closeAll(PreparedStatement pstmt, Connection con) throws SQLException {
+        if (pstmt != null)
+            pstmt.close();
+        if (con != null)
+            con.close();
+    }
+    public static void closeAll(ResultSet rs, PreparedStatement pstmt, Connection con) throws SQLException {
+        if (rs != null)
+            rs.close();
+        closeAll(pstmt, con);
+    }
+    
 }
