--- nest_server_0616/src/org/kosa/nest/common/DatabaseUtil.java
@@ -1,27 +1,0 @@
-package org.kosa.nest.common;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-public class DatabaseUtil {
-	public static Connection getConnection() throws SQLException {
-		return DriverManager.getConnection(DbConfig.URL, DbConfig.USER, DbConfig.PASS);
-	}
-
-	public static void closeAll(PreparedStatement pstmt, Connection con) throws SQLException {
-		if (pstmt != null)
-			pstmt.close();
-		if (con != null)
-			con.close();
-	}
-	public static void closeAll(ResultSet rs, PreparedStatement pstmt, Connection con) throws SQLException {
-		if (rs != null)
-			rs.close();
-		closeAll(pstmt, con);
-	}
-	
-	
-}
